+++ lib/controllers/root-controller.js
@@ -16,7 +16,7 @@ import CredentialDialog from '../views/credential-dialog';
 import Commands, {Command} from '../atom/commands';
 import GitTimingsView from '../views/git-timings-view';
 import GithubTabController from './github-tab-controller';
-import FilePatchController from './file-patch-controller';
+import FilePatchItem from '../items/file-patch-item';
 import IssueishPaneItem from '../items/issueish-pane-item';
 import GitTabItem from '../items/git-tab-item';
 import StatusBarTileController from './status-bar-tile-controller';
@@ -318,24 +318,14 @@ export default class RootController extends React.Component {
       <Fragment>
         <PaneItem
           workspace={this.props.workspace}
-          uriPattern={FilePatchController.uriPattern}>
+          uriPattern={FilePatchItem.uriPattern}>
           {({itemHolder, params}) => (
-            <FilePatchController
+            <FilePatchItem
               ref={itemHolder.setter}
-              deserializers={this.props.deserializers}
-              commandRegistry={this.props.commandRegistry}
+              repository={this.props.repository}
               tooltips={this.props.tooltips}
-              switchboard={this.props.switchboard}
-              getRepositoryForWorkdir={this.getRepositoryForWorkdir}
-              workingDirectoryPath={params.workdir}
-              filePath={path.join(...params.relpath)}
-              initialStagingStatus={params.stagingStatus}
-              didSurfaceFile={this.surfaceFromFileAtPath}
-              didDiveIntoFilePath={this.diveIntoFilePatchForPath}
-              quietlySelectItem={this.quietlySelectItem}
-              openFiles={this.openFiles}
-              discardLines={this.discardLines}
-              undoLastDiscard={this.undoLastDiscard}
+              relPath={path.join(...params.relPath)}
+              stagingStatus={params.stagingStatus}
             />
           )}
         </PaneItem>
@@ -527,7 +517,7 @@ export default class RootController extends React.Component {
       }
       const lineNum = editor.getCursorBufferPosition().row + 1;
       const filePatchItem = await this.props.workspace.open(
-        FilePatchController.buildURI(filePath, repoPath, stagingStatus),
+        FilePatchItem.buildURI(filePath, repoPath, stagingStatus),
         {pending: true, activatePane: true, activateItem: true},
       );
       await filePatchItem.getRealItemPromise();
